Refactor: Separate index operations into value normalization and query modules
Split index_operations.rs (814 lines) into focused modules for better
separation of concerns and code organization:

**New Module Structure:**
- value_normalization.rs: normalize_for_comparison() for canonical value forms
- range_bounds.rs: calculate_next_value(), try_increment_sqlvalue(), smart_increment_value()
- point_lookup.rs: get(), contains_key(), multi_lookup(), iter(), values()
- range_scan.rs: range_scan() for range query operations
- prefix_match.rs: prefix_multi_lookup() for multi-column index prefix matching
- index_operations.rs: Public API re-exports and documentation

**Benefits:**
- Clear separation: Value handling separate from query logic
- Better testability: Each module tested independently with own test suite
- Improved reusability: Normalization logic available to other modules
- Enhanced maintainability: Related functionality grouped together
- Type safety: Foundation for newtype wrappers for normalized values

**Testing:**
- All 194 existing tests pass (1 pre-existing failure in unrelated page module)
- Tests moved to respective modules
- No functional changes - pure refactoring

Closes #2187
