+++ crates/vibesql-storage/src/database/indexes/point_lookup.rs
@@ -1,0 +1,145 @@
+// ============================================================================
+// Point Lookup - Single-value equality operations
+// ============================================================================
+
+use vibesql_types::SqlValue;
+
+use super::index_metadata::{acquire_btree_lock, IndexData};
+use super::value_normalization::normalize_for_comparison;
+
+impl IndexData {
+    /// Lookup exact key in the index
+    ///
+    /// # Arguments
+    /// * `key` - Key to look up
+    ///
+    /// # Returns
+    /// Reference to vector of row indices if key exists, None otherwise
+    ///
+    /// # Note
+    /// This is the primary point-lookup API for index queries.
+    pub fn get(&self, key: &[SqlValue]) -> Option<&Vec<usize>> {
+        match self {
+            IndexData::InMemory { data } => data.get(key),
+            IndexData::DiskBacked { .. } => {
+                // TODO: Implement when DiskBacked is active
+                unimplemented!("DiskBacked lookup not yet implemented")
+            }
+        }
+    }
+
+    /// Check if a key exists in the index
+    ///
+    /// # Arguments
+    /// * `key` - Key to check
+    ///
+    /// # Returns
+    /// true if key exists, false otherwise
+    ///
+    /// # Note
+    /// Used primarily for UNIQUE constraint validation.
+    pub fn contains_key(&self, key: &[SqlValue]) -> bool {
+        match self {
+            IndexData::InMemory { data } => data.contains_key(key),
+            IndexData::DiskBacked { .. } => {
+                // TODO: Implement when DiskBacked is active
+                unimplemented!("DiskBacked contains_key not yet implemented")
+            }
+        }
+    }
+
+    /// Lookup multiple values in the index (for IN predicates)
+    ///
+    /// # Arguments
+    /// * `values` - List of values to look up
+    ///
+    /// # Returns
+    /// Vector of row indices that match any of the values
+    pub fn multi_lookup(&self, values: &[SqlValue]) -> Vec<usize> {
+        match self {
+            IndexData::InMemory { data } => {
+                // Deduplicate values to avoid returning duplicate rows
+                // For example, WHERE a IN (10, 10, 20) should only look up 10 once
+                let mut unique_values: Vec<&SqlValue> = values.iter().collect();
+                unique_values.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
+                unique_values.dedup();
+
+                let mut matching_row_indices = Vec::new();
+
+                for value in unique_values {
+                    // Normalize value for consistent lookup (matches insertion-time normalization)
+                    let normalized_value = normalize_for_comparison(value);
+                    let search_key = vec![normalized_value];
+                    if let Some(row_indices) = data.get(&search_key) {
+                        matching_row_indices.extend(row_indices);
+                    }
+                }
+
+                // Return row indices in the order they were collected from BTreeMap
+                // For IN predicates, we collect results for each value in the order
+                // specified. We should NOT sort by row index as that would destroy
+                // the semantic ordering of the results.
+                matching_row_indices
+            }
+            IndexData::DiskBacked { btree, .. } => {
+                // Deduplicate values to avoid returning duplicate rows
+                let mut unique_values: Vec<&SqlValue> = values.iter().collect();
+                unique_values.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
+                unique_values.dedup();
+
+                // Normalize values for consistent lookup (matches insertion-time normalization)
+                // Convert SqlValue values to Key (Vec<SqlValue>) format
+                let keys: Vec<Vec<SqlValue>> = unique_values
+                    .iter()
+                    .map(|v| vec![normalize_for_comparison(v)])
+                    .collect();
+
+                // Safely acquire lock and call BTreeIndex::multi_lookup
+                match acquire_btree_lock(btree) {
+                    Ok(guard) => guard.multi_lookup(&keys).unwrap_or_else(|_| vec![]),
+                    Err(e) => {
+                        // Log error and return empty result set
+                        log::warn!("BTreeIndex lock acquisition failed in multi_lookup: {}", e);
+                        vec![]
+                    }
+                }
+            }
+        }
+    }
+
+    /// Get an iterator over all key-value pairs in the index
+    ///
+    /// # Returns
+    /// Iterator yielding references to (key, row_indices) pairs
+    ///
+    /// # Note
+    /// For in-memory indexes, iteration is in sorted key order (BTreeMap ordering).
+    /// This method enables index scanning operations without exposing internal data structures.
+    pub fn iter(&self) -> Box<dyn Iterator<Item = (&Vec<SqlValue>, &Vec<usize>)> + '_> {
+        match self {
+            IndexData::InMemory { data } => Box::new(data.iter()),
+            IndexData::DiskBacked { .. } => {
+                // TODO: Implement when DiskBacked is active
+                unimplemented!("DiskBacked iteration not yet implemented")
+            }
+        }
+    }
+
+    /// Get an iterator over all row index vectors in the index
+    ///
+    /// # Returns
+    /// Iterator yielding references to row index vectors
+    ///
+    /// # Note
+    /// This method is used for full index scans where we need all row indices
+    /// regardless of the key values.
+    pub fn values(&self) -> Box<dyn Iterator<Item = &Vec<usize>> + '_> {
+        match self {
+            IndexData::InMemory { data } => Box::new(data.values()),
+            IndexData::DiskBacked { .. } => {
+                // TODO: Implement when DiskBacked is active
+                unimplemented!("DiskBacked values iteration not yet implemented")
+            }
+        }
+    }
+}
